requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
branca==0.6.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
customtkinter==5.0.5
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
folium==0.14.0
fqdn==1.5.1
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
lamini==0.0.16
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nodeenv==1.7.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
openapi-codec==1.3.2
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.6.2
prisma==0.8.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pyrsistent==0.19.3
pytesseract==0.3.10
python-configuration==0.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pytz==2023.3
pywin32==305
pywinpty==2.0.10
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomlkit==0.11.8
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
widgetsnbextension==4.0.5
winstats==0.80
